organize js
